Renamed Screen.setup and App.setup methods to build. Moved screen.py to root dir
